Update shipping calc to return a double instead of setting inline value
As a remote EJB, ShippingService will only receive a copy of the
ShoppingCart object, not the object itself. Any mutator calls will have
no effect. To work around this, we will have the ShippingService return
the cost and have the ShoppingCartService handle the ShoppingCart object
(which would be the proper way to handle these services anyway).